requirements.txt
django>=3.0,<4.0
redis
celery
djangorestframework
django-celery-beat
django-filter
psycopg2
drf_yasg
